+++ app/Filament/Resources/AirlineResource/RelationManagers/StandsRelationManager.php
@@ -2,11 +2,14 @@
 
 namespace App\Filament\Resources\AirlineResource\RelationManagers;
 
+use App\Filament\Helpers\SelectOptions;
 use App\Filament\Resources\Pages\LimitsTableRecordListingOptions;
 use App\Filament\Resources\TranslatesStrings;
+use App\Models\Aircraft\Aircraft;
 use App\Models\Stand\Stand;
 use Carbon\Carbon;
 use Closure;
+use Filament\Forms\Components\Select;
 use Filament\Forms\Components\TextInput;
 use Filament\Forms\Components\TimePicker;
 use Filament\Resources\RelationManagers\RelationManager;
@@ -37,16 +40,24 @@ class StandsRelationManager extends RelationManager
             ->columns([
                 Tables\Columns\TextColumn::make('stand_id')
                     ->formatStateUsing(fn (Stand $record) => $record->airfieldIdentifier)
-                    ->label(self::translateTablePath('columns.terminal'))
+                    ->label(self::translateTablePath('columns.stand'))
+                    ->sortable()
+                    ->searchable(),
+                Tables\Columns\TextColumn::make('aircraft_id')
+                    ->label(self::translateTablePath('columns.aircraft'))
+                    ->formatStateUsing(fn (?int $state) => isset($state) ? Aircraft::find($state)->code : '')
                     ->sortable()
                     ->searchable(),
                 Tables\Columns\TextColumn::make('destination')
                     ->label(self::translateTablePath('columns.destination'))
                     ->default(self::DEFAULT_COLUMN_VALUE)
                     ->sortable(),
-                Tables\Columns\TextColumn::make('callsign_slug')
+                Tables\Columns\TextColumn::make('callsign')
                     ->default(self::DEFAULT_COLUMN_VALUE)
                     ->label(self::translateTablePath('columns.callsign')),
+                Tables\Columns\TextColumn::make('callsign_slug')
+                    ->default(self::DEFAULT_COLUMN_VALUE)
+                    ->label(self::translateTablePath('columns.callsign_slug')),
                 Tables\Columns\TextColumn::make('priority')
                     ->default(self::DEFAULT_COLUMN_VALUE)
                     ->label(self::translateTablePath('columns.priority')),
@@ -62,13 +73,22 @@ class StandsRelationManager extends RelationManager
                             ->getRecordSelect()
                             ->label(self::translateFormPath('icao.label'))
                             ->required(),
+                        Select::make('aircraft_id')
+                            ->options(SelectOptions::aircraftTypes())
+                            ->searchable()
+                            ->label(self::translateFormPath('aircraft.label'))
+                            ->helperText(self::translateFormPath('aircraft.helper')),
                         TextInput::make('destination')
                             ->label(self::translateFormPath('destination.label'))
                             ->helperText(self::translateFormPath('destination.helper'))
                             ->maxLength(4),
-                        TextInput::make('callsign_slug')
+                        TextInput::make('callsign')
                             ->label(self::translateFormPath('callsign.label'))
                             ->helperText(self::translateFormPath('callsign.helper'))
+                            ->maxLength(4),
+                        TextInput::make('callsign_slug')
+                            ->label(self::translateFormPath('callsign_slug.label'))
+                            ->helperText(self::translateFormPath('callsign_slug.helper'))
                             ->maxLength(4),
                         TextInput::make('priority')
                             ->label(self::translateFormPath('priority.label'))