+++ lang/en/stands/form.php
@@ -52,12 +52,20 @@ return [
         'icao' => [
             'label' => 'ICAO Code',
         ],
+        'aircraft' => [
+            'label' => 'Aircraft Type',
+            'helper' => 'The aircraft type that this airline will use when using this stand.'
+        ],
         'destination' => [
             'label' => 'Destination',
             'helper' => 'The destination used for this allocation. Can be partial matches, for example, "EGJ".'
         ],
         'callsign' => [
-            'label' => 'Callsign Slug',
+            'label' => 'Callsign',
+            'helper' => 'The part of the callsign after the ICAO code for this allocation. Must be an exact match.'
+        ],
+        'callsign_slug' => [
+            'label' => 'Partial Callsign',
             'helper' => 'The part of the callsign after the ICAO code for this allocation. Can be partial matches.'
         ],
         'priority' => [